Index.php added
Testing if GitHub will allow index.php as homepage, rather than
index.html. The tutorials all suggest using index.html because that's
what their scripts are designed to recognize, but if they are using
anything like Apache, it should be able to run the php code in the page
before serving it to others.